ARUHA-1462 Move FeatureToggleService and subsclasses to service package
The clue is in the name. It's a service.
