+++ src/main/java/org/zalando/nakadi/config/RepositoriesConfig.java
@@ -15,8 +15,8 @@ import org.zalando.nakadi.repository.kafka.KafkaConfig;
 import org.zalando.nakadi.repository.zookeeper.ZooKeeperHolder;
 import org.zalando.nakadi.repository.zookeeper.ZookeeperConfig;
 import org.zalando.nakadi.service.timeline.TimelineSync;
-import org.zalando.nakadi.util.FeatureToggleService;
-import org.zalando.nakadi.util.FeatureToggleServiceZk;
+import org.zalando.nakadi.service.FeatureToggleService;
+import org.zalando.nakadi.service.FeatureToggleServiceZk;
 import org.zalando.nakadi.validation.EventBodyMustRespectSchema;
 import org.zalando.nakadi.validation.EventMetadataValidationStrategy;
 import org.zalando.nakadi.validation.JsonSchemaEnrichment;
