+++ src/main/java/org/zalando/nakadi/service/ClosedConnectionsCrutch.java
@@ -14,7 +14,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import org.zalando.nakadi.util.FeatureToggleService;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.BufferedReader;
