+++ src/main/java/org/zalando/nakadi/service/EventStreamWriterProvider.java
@@ -3,7 +3,6 @@ package org.zalando.nakadi.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
-import org.zalando.nakadi.util.FeatureToggleService;
 
 @Component
 public class EventStreamWriterProvider {
