+++ src/main/java/org/zalando/nakadi/service/NakadiKpiPublisher.java
@@ -7,7 +7,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.zalando.nakadi.util.FeatureToggleService;
 
 import java.security.MessageDigest;
 import java.util.function.Supplier;
