+++ src/main/java/org/zalando/nakadi/service/StorageService.java
@@ -21,7 +21,6 @@ import org.zalando.nakadi.exceptions.runtime.RepositoryProblemException;
 import org.zalando.nakadi.exceptions.runtime.StorageIsUsedException;
 import org.zalando.nakadi.repository.db.StorageDbRepository;
 import org.zalando.nakadi.repository.zookeeper.ZooKeeperHolder;
-import org.zalando.nakadi.util.FeatureToggleService;
 import org.zalando.problem.Problem;
 
 import javax.annotation.PostConstruct;
