+++ src/test/java/org/zalando/nakadi/service/EventTypeServiceTest.java
@@ -21,7 +21,6 @@ import org.zalando.nakadi.repository.db.SubscriptionDbRepository;
 import org.zalando.nakadi.repository.kafka.PartitionsCalculator;
 import org.zalando.nakadi.service.timeline.TimelineService;
 import org.zalando.nakadi.service.timeline.TimelineSync;
-import org.zalando.nakadi.util.FeatureToggleService;
 import org.zalando.nakadi.validation.SchemaEvolutionService;
 
 import java.util.ArrayList;
