+++ src/test/java/org/zalando/nakadi/service/StorageServiceTest.java
@@ -9,7 +9,6 @@ import org.zalando.nakadi.exceptions.runtime.NoStorageException;
 import org.zalando.nakadi.exceptions.runtime.StorageIsUsedException;
 import org.zalando.nakadi.repository.db.StorageDbRepository;
 import org.zalando.nakadi.repository.zookeeper.ZooKeeperHolder;
-import org.zalando.nakadi.util.FeatureToggleService;
 import org.zalando.nakadi.utils.TestUtils;
 
 import static org.hamcrest.CoreMatchers.equalTo;
